+++ apps/web/app/(dashboard)/layout.tsx
@@ -1,26 +1,9 @@
 "use client";
 
-import { Search } from "lucide-react";
-import { SidebarMenuButton } from "@multica/ui/components/ui/sidebar";
 import { DashboardLayout } from "@multica/views/layout";
 import { MulticaIcon } from "@multica/ui/components/common/multica-icon";
-import { SearchCommand, useSearchStore } from "@/features/search";
+import { SearchCommand, SearchTrigger } from "@multica/views/search";
 import { ChatFab, ChatWindow } from "@/features/chat";
-
-function SearchTrigger() {
-  return (
-    <SidebarMenuButton
-      className="text-muted-foreground"
-      onClick={() => useSearchStore.getState().setOpen(true)}
-    >
-      <Search />
-      <span>Search...</span>
-      <kbd className="pointer-events-none ml-auto inline-flex h-5 select-none items-center gap-0.5 rounded border bg-muted px-1.5 font-mono text-[10px] font-medium text-muted-foreground">
-        <span className="text-xs">⌘</span>K
-      </kbd>
-    </SidebarMenuButton>
-  );
-}
 
 export default function Layout({ children }: { children: React.ReactNode }) {
   return (
