--- apps/web/features/search/index.ts
@@ -1,2 +1,0 @@
-export { SearchCommand } from "./components/search-command";
-export { useSearchStore } from "./stores/search-store";
