+++ packages/views/package.json
@@ -29,11 +29,13 @@
     "./workspace/workspace-avatar": "./workspace/workspace-avatar.tsx",
     "./layout": "./layout/index.ts",
     "./auth": "./auth/index.ts",
+    "./search": "./search/index.ts",
     "./settings": "./settings/index.ts"
   },
   "dependencies": {
     "@multica/core": "workspace:*",
     "@multica/ui": "workspace:*",
+    "cmdk": "^1.1.1",
     "@base-ui/react": "^1.3.0",
     "@dnd-kit/core": "^6.3.1",
     "@dnd-kit/sortable": "^10.0.0",
