+++ packages/views/search/index.ts
@@ -1,0 +1,3 @@
+export { SearchCommand } from "./search-command";
+export { SearchTrigger } from "./search-trigger";
+export { useSearchStore } from "./search-store";
